fix(fossid-webapp): Support multiple source location line ranges
FossID can match several line ranges of the source file to a snippet.
Currently, the `TextLocation` used can carry only a single line range so
further line ranges are ignored and one added to 'additionalData' property
of the snippet.

In a future commit, snippets will be grouped by source locations therefore
multiple line ranges for the source location has to be supported.

Fixes #7045.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>