fix(fossid-webapp): Group snippets by source file matching lines
All snippets matching the same source file location, line ranges included,
are now grouped in the snippet model.
This commit contains the model change and the support in the FossID
scanner. ScanOSS support has not been added but the current implementation
is still functional.

This is in preparation for the "snippet choice" feature that will be added
in a future commit.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>