feat(fossid-webapp): Add license findings from files marked as identified
When a file has been marked as identified in current run or a previous one,
it is not pending anymore thus its snippets won't be listed.
The license however and its source location are attached to the file in
FossID to the identification and a comment. This information is loaded to
create a proper license finding out of it.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>