feat(fossid-webapp): Put unmatched identified files back to 'pending'
When one or several source locations of a 'marked as identified' file have
been deleted from the .ort.yml, put again the file to the pending state.
Its snippets will be listed as for all pending files.
Please note that the logic is not the same if one or all the locations have
been removed. If only one location has been removed, the comment in the
marked as identified file is used to compare the location counts. If all
the locations have be removed from the .ort.yml, just having the file
marked as identified is already the tell.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.io>