fix(fossid-webapp): Add a version check in `waitDownloadComplete`
Previous versions of FossID could stay blocked in the state NOT_FINISHED
while downloading the repository.  Therefore, during the call to
`waitDownloadComplete`, a workaround that checks Git output was put in
place.
However, with the newest version of FossID (23.3.0), trying to call
"trigger scan" when the download is not finished could lead to an error
"Downloading scan files from the git repository is NOT FINISHED. Please
wait a little and click rescan button.".
Therefore, this commit adds a version check to the workaround.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.com>
(cherry picked from commit 3355b85803936a01ba706fffbad9ab3978b44ffe)